requirements.txt
jsonargparse==4.33.2
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
regex==2024.9.11
datasets==2.15.0
konlpy==0.6.0
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
pymilvus[model]==2.4.8
rank_bm25==0.2.2
scikit_learn==1.5.2
seaborn==0.13.2
torch==2.0.1
tqdm==4.66.5
transformers==4.45.2
sentence-transformers==3.0.1
evaluate==0.4.3
nltk==3.9.1
peft==0.13.2
python_box==7.2.0
trl==0.11.4
openai==1.52.2
wandb
